+++ packages/druxt/test/components/DruxtModule.test.js
@@ -13,6 +13,7 @@ describe('DruxtModule component', () => {
     mocks = {
       $createElement: jest.fn(),
       $fetchState: { pending: false },
+      $options: { druxt: {} },
       $nuxt: {
         context: {
           isDev: false,
@@ -133,7 +134,10 @@ describe('DruxtModule component', () => {
       extends: DruxtModule,
       druxt: {
         componentOptions: () => ([['wrapper']]),
+        async fetchConfig() {},
+        async fetchData(settings) {},
         propsData: () => ({ foo: 'bar' }),
+        settings: ({}, settings) => ({ ...settings, custom: true }),
         slots: (h) => ({ default: () => h('div', ['test'] )}),
       }
     }
@@ -148,7 +152,7 @@ describe('DruxtModule component', () => {
       options: ['CustomModuleWrapper'],
       props: { foo: 'bar' },
       propsData: { foo: 'bar' },
-      settings: { foo: 'bar' },
+      settings: { foo: 'bar', custom: true },
       slots: ['default'],
     })
 
@@ -196,11 +200,19 @@ describe('DruxtModule component', () => {
   })
 
   test('custom default slot', async () => {
+    const customModule = {
+      extends: DruxtModule,
+      druxt: {}
+    }
+
     const scopedSlots = { default: jest.fn() }
-    const wrapper = mount(DruxtModule, { localVue, mocks, scopedSlots })
+    const wrapper = mount(customModule, { localVue, mocks, scopedSlots })
     await wrapper.vm.$options.fetch.call(wrapper.vm)
 
     wrapper.vm.getScopedSlots().default.call(wrapper.vm)
     expect(scopedSlots.default).toHaveBeenCalled()
+
+    await wrapper.setProps({ wrapper: true })
+    await wrapper.vm.$options.fetch.call(wrapper.vm)
   })
 })