+++ packages/entity/src/components/DruxtEntity.vue
@@ -40,6 +40,27 @@ export default {
     },
 
     /**
+     * Module settings object.
+     *
+     * @type {ModuleSettings}
+     * @default {}
+     */
+    settings: {
+      type: Object,
+      default: () => ({}),
+    },
+
+    /**
+     * Drupal display schema type, 'view' or 'form'.
+     *
+     * @type {('view'|'form')}
+     */
+    schemaType: {
+      type: String,
+      default: undefined,
+    },
+
+    /**
      * JSON:API resource type.
      *
      * @type {string}
@@ -58,16 +79,6 @@ export default {
       type: [Boolean, String],
       default: false
     },
-
-    /**
-     * Drupal display schema type, 'view' or 'form'.
-     *
-     * @type {('view'|'form')}
-     */
-    schemaType: {
-      type: String,
-      default: undefined,
-    },
   },
 
   /**
@@ -78,54 +89,12 @@ export default {
     model: {
       attributes: {},
       relationships: {},
+      included: undefined,
       type,
       ...value,
     },
     schema: null,
   }),
-
-  /**
-   * Nuxt.js fetch method.
-   */
-  async fetch() {
-    // Fetch Schema.
-    try {
-      this.schema = await this.getSchema({
-        resourceType: this.type,
-        mode: this.mode,
-        schemaType: this.schemaType || 'view',
-      })
-    } catch(e) {
-      // TODO: Handle error
-    }
-
-    // Build wrapper component object.
-    const options = this.getModuleComponents()
-    let component = {
-      is: (((options.filter(o => o.global) || [])[0] || {}).name || 'DruxtWrapper'),
-      options: options.map(o => o.name) || [],
-    }
-
-    // Get wrapper component data to merge with module settings.
-    const wrapperData = await this.getWrapperData(component.is)
-    component.settings = merge(this.$druxt.settings.entity || {}, wrapperData.druxt || {}, { arrayMerge: (dest, src) => src })
-
-    // Fetch Entity resource.
-    if (this.uuid && !this.value) {
-      const query = this.getQuery(component.settings)
-      const entity = (await this.getResource({ type: this.type, id: this.uuid, query })).data
-      this.model = JSON.parse(JSON.stringify(entity || {}))
-    }
-
-    // Get scoped slots.
-    component.slots = Object.keys(this.getScopedSlots())
-
-    // Build wrapper component propsData.
-    component = { ...component, ...this.getModulePropsData(wrapperData.props) }
-
-    // Set component data.
-    this.component = component
-  },
 
   fetchKey(getCounter) {
     const parts = ['DruxtEntity', this.type, this.uuid, this.mode, this.schemaType].filter((o) => o)
@@ -205,15 +174,42 @@ export default {
     getQuery(settings) {
       const query = new DrupalJsonApiParams()
 
-      // Build fields list.
-      const fields = (settings.query || {}).schema
-        ? [...((this.schema || {}).fields || []).map((o) => o.id), ...((settings.query || {}).fields || [])]
-        : ((settings.query || {}).fields || [])
-      if (fields.length) {
-        query.addFields(this.type, fields)
-        return query
+      // Add includes.
+      if (settings.query.include && Array.isArray(settings.query.include)) {
+        query.addInclude(settings.query.include)
       }
-      return false
+
+      let rootFields = []
+      if (settings.query.fields && Array.isArray(settings.query.fields)) {
+        // If the first item is a string, this is a root level filter array.
+        if (settings.query.fields.length === 0 || typeof settings.query.fields[0] === 'string' || typeof settings.query.fields[0] === 'undefined') {
+          rootFields = settings.query.fields
+        // Otherwise this is an array structure field map.
+        } else if (Array.isArray(settings.query.fields[0])) {
+          rootFields = settings.query.fields.find((a) => a.length === 0 || typeof [...a].pop() === 'string' || typeof [...a].pop() === 'undefined') || []
+
+          // Apply included field mapping.
+          const fieldMaps = settings.query.fields.filter(
+            (a) => a.length == 2 && typeof a[0] === 'string' && Array.isArray(a[1])
+          )
+          for (const fieldMap of fieldMaps) {
+            query.addFields(fieldMap[0], fieldMap[1])
+          }
+        }
+      }
+
+      // If Schema mode, generate list including schema fields and explicitly
+      // defined fields.
+      if (settings.query.schema) {
+        rootFields = [...((this.schema || {}).fields || []).map((o) => o.id), ...rootFields]
+      }
+
+      // Apply root fields filter.
+      if (rootFields.length) {
+        query.addFields(this.type, rootFields)
+      }
+
+      return query
     },
 
     /**
@@ -257,12 +253,55 @@ export default {
     ]),
 
     /**
+     * Fetch Schema.
+     */
+    async fetchConfig() {
+      try {
+        this.schema = await this.getSchema({
+          resourceType: this.type,
+          mode: this.mode,
+          schemaType: this.schemaType || 'view',
+        })
+      } catch(e) {
+        // TODO: Handle error
+      }
+    },
+
+    /**
+     * Fetch Entity resource.
+     */
+    async fetchData(settings) {
+      if (this.uuid && !this.value) {
+        const query = this.getQuery(settings)
+        const resource = await this.getResource({ type: this.type, id: this.uuid, query })
+        const entity = resource.data || {}
+        entity.included = resource.included
+        this.model = JSON.parse(JSON.stringify(entity || {}))
+      }
+    },
+
+    /**
      * Provides propsData for the DruxtWrapper.
      *
      * @param {object} context - The module component ViewModel.
      * @returns {PropsData}
      */
     propsData: ({ fields, model, schema }) => ({ entity: model, fields, schema, value: model }),
+
+    /**
+     * Component settings.
+     */
+    settings: ({ $druxt, settings }, wrapperSettings) => {
+      // Start with the `nuxt.config.js` `druxt.settings.entity` settings and
+      // merge the Wrapper component settings on top.
+      let mergedSettings = merge($druxt.settings.entity || {}, wrapperSettings, { arrayMerge: (dest, src) => src })
+      // Merge the DruxtEntity component `settings` property on top.
+      mergedSettings = merge(mergedSettings || {}, settings, { arrayMerge: (dest, src) => src })
+      // Currently only returning the query settings.
+      return {
+        query: mergedSettings.query || {},
+      }
+    },
 
     /**
      * Provides the scoped slots object for the Module render function.
@@ -344,33 +383,46 @@ export default {
  */
 
 /**
- * Provides settings for the Entity module, via the `nuxt.config.js` `druxt.entity`
- * or the Wrapper component `druxt` object.
+ * Provides settings for the Entity module, via the `nuxt.config.js` `druxt.entity`,
+ * the Wrapper component `druxt` object or the DruxtEntity component `settings`
+ * property.
  *
  * @typedef {object} ModuleSettings
- * @param {string[]} fields - An array of fields to filter from the JSON:API Resource.
- * @param {boolean} schema - Whether to automatically detect fields to filter, per the Display mode.
+ * @param {object} query - Entity Query settings:
+ * @param {(string[]|array[])} query.fields - An array or arrays of fields to filter from the JSON:API Resources.
+ * @param {string[]} query.include - An array of relationships to include alongside the JSON:API Resource.
+ * @param {boolean} query.schema - Whether to automatically detect fields to filter, per the Display mode.
  *
- * @example @lang js
- * {
- *   fields: [],
- *   schema: false,
- * }
- *
- * @example @lang vue
+ * @example <caption>DruxtEntity Wrapper component</caption> @lang vue
  * <script>
  * export default {
  *   druxt: {
  *     query: {
- *       fields: ['title']
+ *       fields: [['title'], ['user--user', ['display_name']]],
+ *       include: ['uid']
  *       schema: true,
  *     },
  *   }
  * }
+ *
+ * @example <caption>DruxtEntity component with settings</caption> @lang vue
+ * <template>
+ *   <DruxtEntity
+ *     type="node--article"
+ *     :uuid="uuid"
+ *     :settings="{
+ *       query: {
+ *         fields: [['title'], ['user--user', ['display_name']]],
+ *         include: ['uid']
+ *         schema: true,
+ *       }
+ *     }"
+ *   />
+ * </template>
  */
 
 /**
- * Provides propsData for use in the Wrapper component.
+ * Provides property data for use in the Wrapper component.
  *
  * @typedef {object} PropsData
  * @param {object} entity - The Drupal Entity JSON:API resource data.