+++ packages/menu/src/nuxtStorybook.js
@@ -14,7 +14,7 @@ parts
 export default async function ({ stories }) {
   const { addTemplate, options } = this
 
-  const druxt = new DruxtClient(options.druxt.baseUrl, options.druxt)
+  const druxt = new DruxtClient(options.druxt.baseUrl, { ...options.druxt, proxy: { api: false } })
 
   const resourceType = 'menu--menu'
   const query = new DrupalJsonApiParams()