+++ packages/schema/src/mixins/schema.js
@@ -7,30 +7,18 @@ import { mapActions } from 'vuex'
  *
  * @example @lang vue
  * <script>
- * // Import mixin.
  * import { DruxtSchemaMixin } from 'druxt-schema'
- *
  * export default {
- *   name: 'CustomComponent',
- *
- *   // Register mixin.
  *   mixins: [DruxtSchemaMixin],
  * }
  * </script>
- *
- * @example @lang vue
- * <!-- Render component with lazy loaded Default View mode Page schema. -->
- * <CustomComponent type="node--page" mode="default" />
- *
- * @todo Add schemaType property to Mixin.
  */
 const DruxtSchemaMixin = {
-  /**
-   * Vue.js Properties.
-   */
+  /** */
   props: {
     /**
      * The Drupal Display mode.
+     *
      * @type {string}
      * @default default
      */
@@ -40,34 +28,45 @@ const DruxtSchemaMixin = {
     },
 
     /**
+     * Drupal display schema type, 'view' or 'form'.
+     *
+     * @type {('view'|'form')}
+     */
+     schemaType: {
+      type: String,
+      default: undefined,
+    },
+
+    /**
      * The JSON:API Resource type.
+     *
      * @type {string}
      */
     type: {
       type: String,
       required: true
-    }
+    },
   },
 
   /**
    * Loads the Schema from the Vuex store.
    */
   async fetch() {
-    this.schema = await this.getSchema({ resourceType: this.type, mode: this.mode })
+    this.schema = await this.getSchema({
+      resourceType: this.type,
+      mode: this.mode,
+      schemaType: this.schemaType || 'view'
+    })
   },
 
   /**
-   * Vue.js Data object.
-   *
    * @property {object} schema - The Drupal Schema data.
    */
   data: () => ({
     schema: false
   }),
 
-  /**
-   * Vue.js methods.
-   */
+  /** */
   methods: {
     /**
      * Maps `druxtSchema/get` Vuex action to `this.getSchema`.