+++ packages/views/CHANGELOG.md
@@ -1,5 +1,20 @@
 # druxt-views
 
+## 0.16.0
+
+### Minor Changes
+
+- e7b1533: Refactored DruxtModule fetch hooks
+
+### Patch Changes
+
+- d80f5d4: Fixed issue with undefined wrapper in DruxtViewsFilter
+- Updated dependencies [025315a]
+  - druxt-entity@0.21.0
+  - druxt-blocks@0.14.0
+  - druxt@0.13.0
+  - druxt-router@0.23.0
+
 ## 0.15.0
 
 ### Minor Changes